+++ src/main/kotlin/com/github/insanusmokrassar/IObjectK/interfaces/CommonIObject.kt
@@ -1,5 +1,3 @@
 package com.github.insanusmokrassar.IObjectK.interfaces
 
 interface CommonIObject<KeyType, ValueType> : IInputObject<KeyType, ValueType>, IOutputObject<KeyType, ValueType>
-
-typealias IObject<T> = CommonIObject<String, T>