+++ src/main/kotlin/com/github/insanusmokrassar/IObjectK/interfaces/IObject.kt
@@ -1,0 +1,3 @@
+package com.github.insanusmokrassar.IObjectK.interfaces
+
+interface IObject<T>: CommonIObject<String, T>