+++ src/main/kotlin/com/github/insanusmokrassar/IObjectK/realisations/ConcurrentSimpleCommonIObject.kt
@@ -2,37 +2,37 @@ package com.github.insanusmokrassar.IObjectK.realisations
 
 import com.github.insanusmokrassar.IObjectK.exceptions.ReadException
 import com.github.insanusmokrassar.IObjectK.exceptions.WriteException
-import com.github.insanusmokrassar.IObjectK.interfaces.IObject
+import com.github.insanusmokrassar.IObjectK.interfaces.CommonIObject
 
-class ConcurrentSimpleIObject : SimpleIObject {
+class ConcurrentSimpleCommonIObject<K, V> : SimpleCommonIObject<K, V> {
 
-    constructor(from: Map<String, Any>) : super(from)
+    constructor(from: Map<K, V>) : super(from)
 
-    constructor(from: IObject<Any>) : super(from)
+    constructor(from: CommonIObject<K, V>) : super(from)
 
     constructor() : super()
 
     @Synchronized @Throws(WriteException::class)
-    override fun put(key: String, value: Any) {
+    override fun put(key: K, value: V) {
         super.put(key, value)
     }
 
     @Synchronized @Throws(WriteException::class)
-    override fun putAll(toPutMap: Map<String, Any>) {
+    override fun putAll(toPutMap: Map<K, V>) {
         super.putAll(toPutMap)
     }
 
     @Synchronized @Throws(ReadException::class)
-    override operator fun <T : Any>get(key: String): T {
-        return super.get<T>(key)
+    override operator fun <T : V>get(key: K): T {
+        return super.get(key)
     }
 
     @Synchronized @Throws(WriteException::class)
-    override fun remove(key: String) {
+    override fun remove(key: K) {
         super.remove(key)
     }
 
-    @Synchronized override fun keys(): Set<String> {
+    @Synchronized override fun keys(): Set<K> {
         return super.keys()
     }
 }